Implemented two more directions to render tooltip
Rrose works very well in solving the auto panning problem but when the tooltip is too big to be rendered in north or south direction, it fails to render tooltip within the viewport. So I implemented two new directions namely east and west apart from the already existing (n, ne, nw, s, se, sw). The tooltip is fixed along the y axis and starts from the top of the view port, the tip of the tooltip, moves according to the point of selection along the y axis. The tooltip is rendered east or west based on the space available.